+++ jax/_src/api.py
@@ -2530,7 +2530,7 @@ def pspec_to_sharding(val):
     mesh = get_concrete_mesh()
     if mesh is None:
       raise ValueError(
-          "Please set a mesh via `jax.sharding.set_mesh` if a PartitionSpec is"
+          "Please set a mesh via `jax.set_mesh` if a PartitionSpec is"
           " passed to device_put")
     return NamedSharding(mesh, val)
   return val